[imageMatching] Extract matching code from main_imageMatching.cpp
This allows the code to be reused by other components.
