+++ src/aliceVision/CMakeLists.txt
@@ -17,6 +17,7 @@ if(ALICEVISION_BUILD_SFM)
   add_subdirectory(geometry)
   add_subdirectory(graph)
   add_subdirectory(gpu)
+  add_subdirectory(imageMatching)
   add_subdirectory(keyframe)
   add_subdirectory(linearProgramming)
   add_subdirectory(localization)
