+++ src/aliceVision/imageMatching/CMakeLists.txt
@@ -1,0 +1,16 @@
+# Headers
+set(imageMatching_files_headers
+    ImageMatching.hpp
+)
+
+# Sources
+set(imageMatching_files_sources
+    ImageMatching.cpp
+)
+
+alicevision_add_library(aliceVision_imageMatching
+    SOURCES ${imageMatching_files_headers} ${imageMatching_files_sources}
+    PRIVATE_LINKS
+        aliceVision_image
+        aliceVision_voctree
+)
